Added Debug and Release build configurations Added Macro _assert(expr) to compile only in Debug mode

+++ BuildAndRun.sh
@@ -1,16 +1,25 @@
-#!/bin/bash 
+#!/bin/bash --
+# You can now call this script without or with arguments 
+# ./BuildAndRun.sh BUILD_TYPE
+# to compile in Release (default) or Debug
+
+build_type=${1:-Release}
+
+if [ "$build_type" != "Release" ] && [ "$build_type" != "Debug" ]; then
+    echo "Invalid build type, use Debug or Release. (Build type: $build_type)"
+    exit 1
+fi
 
 source_dir="$PWD"
-build_dir="$source_dir/build"
+build_dir="$source_dir/build/$build_type"
 
-# Check if the build directory exists
 if [ -d "$build_dir" ]; then
     echo "Build directory already exists. Deleting..."
     rm -r "$build_dir"
 fi
 
 mkdir -p "$build_dir"
-cmake -S "$source_dir" -B "$build_dir"
+cmake -S "$source_dir" -B "$build_dir" -DBUILD_TYPE="$build_type"
 cd "$build_dir"
 make && ./SchaakPlezier
 cd "$source_dir"
